+++ AppKit/NSSegmentedControl/NSSegmentItem.m
@@ -43,9 +43,9 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
 }
 
 - (id) description {
-    return [NSString stringWithFormat: @"%@ - %@ (%f) [%d]",
+    return [NSString stringWithFormat: @"%@ - %@ (%f) [%ld]",
                                        [super description], _label, _width,
-                                       _tag];
+                                       (long) _tag];
 }
 
 - (void) dealloc {
